Bugfix for glow direction
Description here:
https://stackoverflow.com/questions/43026748/three-js-how-to-make-objects-with-different-positions-look-at-orthographic-cam
The problem with the existing implementation is that the mesh for the
glow only looks right from one perspective, and in theory that's
supposed to track with where the camera is. Unfortunately, this is not
currently the case, so I changed it to proejct onto the camera's view.

This still has some problems when the camera isn't looking directly at
the glowing mesh, but for an object that's sufficiently far from the
camera and a glow mesh that's sufficiently small, the imperfection is
more or less invisible.