Create actual bugfix from be04e59a234bfca88a4a1c5b474d080f86381472
The bugfix described in commit be04e5 is actually incorrect, because the
camera used by Aframe is already perspective and not orthographic, so
projecting the vector as done there doesn't solve the problem.

The actual bug was that the view vectors were being computed using the
local, rather than world, positions of the glow meshes. As a result, the
mesh view vector was being computed as though the glow mesh was at
(0, 0, 0) instead of its actual coordinate. This commit fixes that, and
now the glowing mesh is oriented correctly.